# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Challenge {
  id: number;
  title: string;
  summary: string;
  category: string;
  location: string;
  status: string;
  impact: string;
  participants: number;
  partners: string[];
  timeLeft: string;
  createdBy: string;
  createdAt: string;
}

export interface ChallengeInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  summary: string;
  category: string;
  location: string;
  impact: string;
  createdBy?: string;
}

export interface JoinChallengeInput {
  role?: string;
}

export interface Partner {
  id: number;
  name: string;
  type: string;
  focus: string;
  location: string;
  openProjects: number;
  initials: string;
  color: string;
}

export interface Summary {
  activeChallenges: number;
  projectsInMotion: number;
  peopleContributing: number;
  partnerOrganizations: number;
  solvedThisYear: number;
}

export interface AiFrameInput {
  /** @minLength 10 */
  rawIdea: string;
  context?: string;
}

export interface AiFrame {
  title: string;
  framing: string;
  outcomes: string[];
  tags: string[];
  suggestedPartners: string[];
}

export type ListChallengesParams = {
category?: string;
status?: string;
};

export type ListPartnersParams = {
type?: string;
};