# lib/api-zod/src/generated/types/aiFrame.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiFrame {
  title: string;
  framing: string;
  outcomes: string[];
  tags: string[];
  suggestedPartners: string[];
}