# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Challenge {
  id: number;
  title: string;
  summary: string;
  category: string;
  location: string;
  status: string;
  impact: string;
  participants: number;
  partners: string[];
  timeLeft: string;
  createdBy: string;
  createdAt: string;
}