# lib/api-zod/src/generated/types/challengeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  summary: string;
  category: string;
  location: string;
  impact: string;
  createdBy?: string;
}