# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Partner {
  id: number;
  name: string;
  type: string;
  focus: string;
  location: string;
  openProjects: number;
  initials: string;
  color: string;
}