# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Summary {
  activeChallenges: number;
  projectsInMotion: number;
  peopleContributing: number;
  partnerOrganizations: number;
  solvedThisYear: number;
}